requirements.txt

django
djangorestframework
django-environ
psycopg2-binary
gunicorn
python-jose